requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.1.7
django-environ==0.10.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
psycopg2==2.9.6
pycparser==2.21
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==2.0.2
wincertstore==0.2
